Added logging API request as a curl command
Added a logging method which builds a curl command pattern based on the
actual params provided in the API request.  This curl command is logged
when the Logger level is set to FINEST.  This can be useful for running
new APIs as curl commands and comparing wrapper made curl commands with
the ones required by Mambu.
Added “static” to some methods in RequestExecutorImpl.  At some point
we may decide to consider moving all the logging stuff into a separate
Logger helper to avoid cluttering the request executor