Allow appKey for JSON requests to be provided with the input JSON string
Currently, the appKey parameter is added to the JSON string (when
required) by the RequestExecutor. This sometimes creates memory issues
as it requires modifying the already generated large strings (e.g. the
JSON for the encoded image file of 5MB).
1. Allow appKey parameter to be added by a service when creating the
JSON string.
2.  Modify RequestExecutor to add appKey when posting JSON requests
only if the apply is not already present in the input JSON string
3. Move addAppKeyValueToJson() method to the ServiceHelper to allow it
be used by services