Updated ServiceHelper to support retrieving specific fields for a generated JSON string from the nested class members
Need to support MBU-11481 (DisbursementDetails)

With the introduction of the DisbursementDetails in a LoanAccount, such
fields as firstRepaymentDate and expectedDisbursementDate were moved
inside the new disbursementDetails member.  The ServiceHelper was
enhanced to be able to go recursively through to all class members when
looking for a required field (to be able to find such fields as
firstRepaymentDate and expectedDisbursementDate in the
disbursementDetails).

The helper was also enhanced to support two modes of generating a JSON
string with a subset of fields: a JSON string preserving the original
structure of the fields or a ‘flat” JSON, where all fields are
specified linear at the top level.
Using the firstRepaymentDate as an example , the flat version JSON
string would be {"loanAccount":{"loanAmount":"1000",
"firstRepaymentDate":"2016-02-23T16:00:00-0800”}}
While the JSON preserving the “original” structure would be
{“loanAccount":{"loanAmount":"1000", “disbursementDetails”:
{”firstRepaymentDate":"2016-02-23T16:00:00-0800”}}}

Note, currently only the ‘flat’ JSON  strings are needed to support 4.0
APIs