Update Loan Create and Loan Update API methods to use LoanAccount instead of LoanAccountExpanded
1.  Deprecated LoanAccountExpanded class
a) There is now an identical JSONLoanAccount in Mambu model library. It
can be used to generate requests and deserialize responses when needed.
b) there is no need for API users to to use internal
LoanAccountExpanded: API users should be able to submit requests to
create and update loan account using LoanAccount objects
Note: this was also requested as part issue #55 (as using
LoanAccountExpanded is inconvenient)

2. Added new methods for create and update loan account APIs using
LoanAccount and returning LoanAccount
3. Deprecated create and update loan account API methods using
LoanAccountExpanded.